+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Changed
+
+- terraform scripts to require no variables
+
 ## [0.13.0] - 2023-11-27
 
 ### Authors